# Conflicts:
#	src/main/resources/quries.properties

+++ src/main/java/com/zenfra/controller/ReportDataController.java
@@ -3,6 +3,9 @@ package com.zenfra.controller;
 import java.io.File;
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.nio.file.FileSystems;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
@@ -13,6 +16,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
+import java.util.stream.Collectors;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -120,7 +124,6 @@ public class ReportDataController {
 	@PostMapping("saveLocalDiscoveryDF")
 	public ResponseEntity<String> saveLocalDiscoveryDF(@RequestParam("siteKey") String siteKey,
 			@RequestParam("sourceType") String sourceType, @RequestBody JSONObject localDiscoveryData) {
-		System.out.println("---------------api entered to add dataframe-----------------------");
 
 		try {
 			if (localDiscoveryData != null && !localDiscoveryData.isEmpty() && siteKey != null && !siteKey.isEmpty()
@@ -152,11 +155,7 @@ public class ReportDataController {
 		System.out.println("---------------api to add default fav view-----------------------" + sourceType + " : "
 				+ siteKey + " : " + userId);
 
-		try {
-			/*
-			 * if(sourceType != null && !sourceType.trim().isEmpty() &&
-			 * sourceType.trim().equalsIgnoreCase("Tanium")) { sourceType="Linux"; }
-			 */
+		try {			 
 
 			try { // remove orient db dataframe
 				String dataframePath = File.separator + "opt" + File.separator + "ZENfra" + File.separator + "Dataframe"
@@ -165,11 +164,30 @@ public class ReportDataController {
 																											// +
 																											// File.separator;
 				File[] directories = new File(dataframePath).listFiles(File::isDirectory);
-				for (File dir : directories) {
-					if (dir.getName().equalsIgnoreCase(sourceType)) {
-						FileSystemUtils.deleteRecursively(dir);
+				if(directories != null)  {
+					for (File dir : directories) {
+						if (dir.getName().equalsIgnoreCase(sourceType)) {
+							FileSystemUtils.deleteRecursively(dir);
+						}
 					}
 				}
+				
+				
+				try { // delete end to end df file for all log folders
+					Path  configFilePath = FileSystems.getDefault().getPath(dataframePath);
+
+				    List<Path> fileWithName = Files.walk(configFilePath)
+				            .filter(s -> s.toFile().getAbsolutePath().toLowerCase().contains("end-to-end")).collect(Collectors.toList());
+				          
+
+				    for (Path name : fileWithName) {
+				    	FileSystemUtils.deleteRecursively(name);
+				    }
+				
+				} catch (Exception e) {
+					// TODO: handle exception
+				} 
+				
 
 			} catch (Exception e) {
 				e.printStackTrace();
